+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.14.1] - 2024-10-09
+### Fixed
+- Fix query for Meta data dependancies [#132](https://github.com/rokwire/content-building-block/issues/132)
+
 ## [1.14.0] - 2024-10-08
 ### Added
 - Meta_data dependancies [#125](https://github.com/rokwire/content-building-block/issues/125)